Update tests to passing
Since I had built out all the tests prior, within a creat-react-app,
getting them to run was somewhat painful.

Things that matter:

CRA's react-scripts are not updated to the version of Jest where
mockResolvedValue(Once) is a part of the mocking capacity of jest.fn()

That means that you are REQUIRED to eject and modify your jest to run on
22.4.4.

You may be noticing that the 22.4.4 is not the latest version.  That is
because Jest has an issue on windows that screws up the test discovery
for the newer version.  Only by sticking to that exact version do we
gain the benefits.

How do I know this?  Pain.  Lots and lots of pain.

Otherwise, I'll need to stabalize the API for the two components before
I add this in to any projects.  However, once it's done, it'll be quite
neat I think.