Update docs and stabalize
I pulled the withThingContext out into utils and made it withContext.
The signature needed to have propName added, so that the context
provides the value, and the propName acts as a key.  This simplifies the
code somewhat, but it doesn't allow the user to update the value
provided without using a new provider.  I think that's fine for the time
being, but we'll see if I ever get any feedback on that.

I've realized that there are some pretty big things that I'll want to do
with the Socket.On functionality.  This really needs to have a wrapper
that is approximately equivalent in function as the Ajax component.
It'll be an interesting challenge to see if I can factor that out into a
super specialized HOC.  I doubt it's reasonable, given the
implementation requirements of each, but eh.

I also made the README the unofficial, official documentation.  I'll
want to move to docgen eventually, but for the time being, I'll just
keep track of things this way.