Update Socket API and tests
This is a relatively big shift, since I needed a more maintainable and
readable way to make the decision about what to render within the
socket's On and Emit components.  The pattern I landed on was to have a
prop signature defined for a given "type" of component.  That is, when
the component is being used to wrap children, there is a signature of
props that is to be expected.  If that signature is met, then the
wrapper is used.  I don't see how I'll grow to hate this, but I'm sure I
will.

The tests required a good update as well, since I wanted to be able to
mock an emit and server response for a few of them, so the addition of
the mock-socket library was a necessity.  I updated a LOT of test code
to be more concise and correctly use matchers.  There was a hang up with
getting jest to work correctly with the mock-socket inits.  This was
fixed by telling jest that I _do_ in fact want to mock out the timers.
Then YOU have to explicitly state when the timers need to be updated.

Overall it's not that bad, but getting this thing rolling in a real way
is important.