Fix WrapWithProps
Explicitly handling the children is something that I think was missing
in the original function.  I believe that this was the source of the bug
in yggdrasil where the Provider was trying to be wrapped by a Socket.On
with implicit passing of the dataProp='value'.  It was happening, but
the cloneElement call appeared to have created a new provider element
and then used the old one as a child.  This murdered the whole point of
the thing since the old provider had an undefined value.  Thus, any
consumer below that node was receiving undefined for the context value.

Super lame.