Remove showLoading prop
This was a redundant property since the presence or absence of the
loader prop is enough to determine whether or not the loader should
display.  I'm open to the argument that having a boolean prop that
determines the display is cleaner on the consumer than having to perform
a check about whether to pass in the property though.  Personally I like
to have the option, but I'm trying to be as concise as possible.