Add displayName to withContext
The rendering of the displayName in react tools in the browser was
Unknown, which is really not very nice :(
So I re-wrote the withContext function to display the standard of
`withContext(${Component.displayName})`

This is the best way I can think of ATM to have the user inject the
context as a wrapper around the components.  I don't know how you would
do that with composition.  It allows the user to use these components in
a more composable way as well, so I think that it out-weighs any
dogmatic holding to standards.