Add pickEvents and createHandler
Two utility functions that were formerly methods in the Socket.Emit
component.  They are, respectively, methods for generating event
handlers for valid react DOM events from props.

pickEvents currently returns an array of strings, but I don't know if I
want it to return that, or a single string.  I can work with it for now,
but I don't know if I'll ever use it to it's full potential.

createHandler may need some work.  As it stands, the emission payload
for the socket event needs to be generated by the client.  There is no
way to have the payload carry event information.  That is, if the client
wants the target <x, y> of the click event to be in the payload, then
they're SOL.  I don't like this, as it isn't how I envisioned the
various use cases.  I'm also not sure how to make that work.

It will likely have to package the emission again.  Perhaps by passing
in a function instead of an object, the runtime decision can be made to
invoke the function with the event in order to generate the payload from
the event data.  Ooooo. I like that!