Update API for domEvents
The createEvent and pickEvent function are now moved to utils, since I
need that functionality in both Socket and Ajax.  The domEvent prop is
no more.  Instead, the event name itself is treated as a flag and used
to generate the event handler dynamically.  I feel like this is MUCH
simpler.  Since the function handling is implicit in the Socket.Emit
verbiage, having to type extra information was just a pain.  I do,
however, need to get the issue with passing event data to the payload
fixed.