Add ability to hook in to DOM event data
I realized today that not having a way to access the DOM event for the
payload of the emission was a serious shortcoming.  So I've created a
way to hook into it.  I'll need to do some SERIOUS updates to the
documentation over this.

Tests are all passing as well.  Cause I rock.