Update Ajax tests for wrapping
The wrapping functionality isn't complete, but it needs to be.  I also
need to minimally introduce the functionality for blasting POSTs out.  I
think that the basic CRUDs are really all that needs support.  GET
should definitely be the default, but GET, READ, POST, DELETE, UPDATE
should all probably be supported for the initial release.