Update Ajax catch
I realized, while peeing (as this is akin to the mighty shower in terms
of thought provocation), that the setState warnings I was seeing in
production were likely due to setting the state in the catch.  I was
correct.  The cancel token throws an error, and performing a check on
that should eliminate that irritating aspect of the library entirely.

Weeeee!