Add integration tests for Ajax
Since the Ajax component is integrated with the axios library, I found
that getting tests that cover that integration is going to be at least a
little important.  Particularly in the case of how the cancellation of
requests works.  It's finicky, but having that visibilty and reasoning
documented is going to be really helpful when the system needs to change
over to the fetch API at some point.