Add update logic to Socket.On
Issue:
Socket.On was failing to update the event on the socket, meaning that if
the client passed in a new event string, the old was still the one
attached to the event.  This is bad news bears.

Solution:
Add a little logic in the shouldComponentUpdate method.  It's what that
lifecycle is FOR.