Add Socket.RequestResponse
This is a component that will allow me to chunk the logic for an emit/on
pair that should behave more like a traditional request.  What I'm
thinking right now is:
1. It tracks emission and receipt of events, allowing you to define what
happens in certain logical cases.  That is, "if event has fired, then
don't receive, otherwise do" kind of thing

2. It allows two different events to be passed, one for the emit, and
another for the receipt.  These should be interpolatable.

3. It should define an error handler, and be based on promises.  That
is, the emit should trigger the client function passed in to receive a
promise that will allow the client to update whatever state they want
according to the state of the promise. This can then be resolved with
the function passed to 'on'.

That's what I've got for now, but I will definitely specify it better as
I continue to write the tests and the full API becomes more apparent.