Remove RequestResponse
It was a silly thing!

I had forgotten that the socket has an optional function argument
designed for a server ACK function call.  This is basically the response
part of a request response.  I just need to have a hook for it in the
props to the Emit component.

Basically it will just accept data as a parameter and that's it.  I
think that should be sufficient, but I'll think it through more as I
finish adding it.