Wire in acknowledge prop on Emit with tests
Basic acknowledge functionality is now completely wired in.  I have a
few more ideas about how to test it, but I think that it is ready to
give a go in the downstream project to see if there's anything that I
haven't thought of.  I _think_ that I haven't done anything that will
break the downstream horribly, if at all.